+++ 华为机试题/动态规划/两种24点(有无括号)的区别.md
@@ -1,0 +1,22 @@
+## 有括号的24点： 
+1. 枚举所有数字排列（4! = 24种）
+2. 枚举所有合法的括号结构（5种二叉运算结构）
+3. 枚举所有运算符组合（4³ = 64种）
+4. 递归构建表达式树（DFS），逐步运算
+5. 判断是否等于 24
+
+---
+## 无括号的24点：
+
+1. 枚举所有数字排列（4! = 24种）
+2. 枚举运算符组合 op1, op2, op3（4³ = 64种）
+3. 按照顺序 (((a op1 b) op2 c) op3 d) 执行运算
+4. 判断结果是否等于 24
+
+**有括号版本**，需要用递归+回溯的方法做深度优先搜索DFS，进行剪枝和扩展
+
+**无括号版本**，则完全转换为全排列问题：
+获取输入的4个数字所有的可能顺序→定义一个搜索函数，因为计算符的顺序固定，所以依次测试i,j,k即可
+因为运算符只有4种加减乘除，所以复杂度为o(4×4×4)。
+
+导入得到的全排列列表，如果符合条件24点，那么返回当前的序列元素，同时记录当前的计算符号。
